short dir edge case

+++ modules/exporter.py
@@ -119,6 +119,8 @@ class Exporter(threading.Thread):
         _, dirs, __ = next(os.walk(DOWNLOAD_PATH))
         for d in dirs:
             _, __, files = next(os.walk(f"{DOWNLOAD_PATH}/{d}"))
+            if "shorts" in d:
+                continue
             if len(files) == 0:
                 os.rmdir(f"{DOWNLOAD_PATH}/{d}")
             else:
